+++ index.html
@@ -452,6 +452,6 @@
         Загружаем...
       </div>
     </template>
-    <script src="/js/main.js"></script>
+    <script src="/js/main.js" type="module"></script>
   </body>
 </html>
